use case refactoring
